Fix the Wallet connection problems
